move paddle/go to go, move go/cclient to go/pserver/cclient, move go/crecordio to go/recordio/c
