+++ go/cmake/CMakeGoInformation.cmake
