+++ go/pserver/cclient/CMakeLists.txt
@@ -1,6 +1,7 @@
 cmake_minimum_required(VERSION 3.0)
 
 get_filename_component(PARENT_DIR ${CMAKE_CURRENT_SOURCE_DIR} DIRECTORY)
+get_filename_component(PARENT_DIR ${PARENT_DIR} DIRECTORY)
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${PARENT_DIR}/cmake")
 
 project(cxx_go C Go)
@@ -8,5 +9,5 @@ project(cxx_go C Go)
 include(golang)
 include(flags)
 
-go_library(recordio STATIC)
+go_library(client STATIC)
 add_subdirectory(test)
