+++ go/recordio/c/register.go
