+++ go/recordio/c/test/CMakeLists.txt
