+++ go/recordio/c/test/test.c
