+++ go/recordio/range_scanner.go
