+++ go/recordio/recordio_internal_test.go
