+++ go/recordio/scanner.go
