+++ go/recordio/writer.go
